MR : Feat: Add Repository of FavoriteMachine, Machine, Question, Vote | JeongYeji => repository

See merge request s09-webmobile3-sub2/S09P12A304!40

+++ BackEnd/ggong/src/main/java/com/a304/ggong/entity/Machine.java
@@ -18,7 +18,7 @@ public class Machine {
 	@Id
 	@GeneratedValue
 	@Column(name = "machine_no")
-	private int machineNo;
+	private Long machineNo;
 
 	@Column(name = "latitude")
 	private int latitude;
